backend/requirements.txt
# Core dependencies
livekit>=0.20.0
livekit-agents[google,silero]>=1.0.0
livekit-plugins-google>=1.0.0
livekit-plugins-silero>=1.0.0

# FastAPI for API server
fastapi==0.115.5
uvicorn[standard]==0.32.1
websockets==14.1
python-multipart==0.0.20

# MCP (Model Context Protocol) support
mcp==1.4.0
pydantic==2.10.3
httpx==0.28.1

# Odoo integration
odoorpc==0.10.1


# Environment and utilities
python-dotenv==1.0.1
aiofiles==24.1.0
pillow==10.3.0

# Noise cancellation and audio
numpy==2.2.0
scipy==1.15.0
